+++ src/main/kotlin/com/nbadal/ktlint/KtlintAnnotator.kt
@@ -20,12 +20,10 @@ class KtlintAnnotator : ExternalAnnotator<PsiFile, List<LintError>>() {
     }
 
     override fun apply(file: PsiFile, errors: List<LintError>, holder: AnnotationHolder) {
-        val doc = file.viewProvider.document!!
         val config = KtlintConfigStorage.instance(file.project)
 
         errors.forEach {
-            // We only get line + col currently, so mark the whole line.
-            val errorRange = TextRange(doc.getLineStartOffset(it.line - 1), doc.getLineEndOffset(it.line - 1))
+            val errorRange = errorTextRange(file, it)
             val message = "${it.detail} (${it.ruleId})"
             val severity = if (config.treatAsErrors) HighlightSeverity.ERROR else HighlightSeverity.WARNING
 
@@ -34,5 +32,16 @@ class KtlintAnnotator : ExternalAnnotator<PsiFile, List<LintError>>() {
                 registerFix(KtlintConfigAction())
             }
         }
+    }
+
+    private fun errorTextRange(file: PsiFile, it: LintError): TextRange {
+        val doc = file.viewProvider.document!!
+        val lineStart = doc.getLineStartOffset(it.line - 1)
+        val errorOffset = lineStart + (it.col - 1)
+
+        // Full line range in case we can't discern the indicated element:
+        val fullLineRange = TextRange(lineStart, doc.getLineEndOffset(it.line - 1))
+
+        return file.findElementAt(errorOffset)?.let { TextRange.from(errorOffset, it.textLength) } ?: fullLineRange
     }
 }